feat: disorder selector in web dashboard + precision phenotype manuscript updates
- Add /api/disorder and /api/disorders endpoints for live disorder application
- Add disorder dropdown + intensity slider to Playground monitor panel
- Update manuscript disorder table with precision parameters (γ_ceil/γ_floor)
- Add "Precision-based disorder modeling" paragraph (Adams 2013 framework)
- Update ASD paragraph with Lawson 2014 uneven precision
- Update schizophrenia paragraph with γ_ceiling and 3x FE reduction
- Update technical report disorder presets section
- Fix MPS float64 dtype error in proprioception.py
- Fix generate_fe_trace.py import path

+++ zebrav2/brain/proprioception.py
@@ -108,11 +108,11 @@ class SpikingProprioception(nn.Module):
         coll_val = 1.0 if self.collision else 0.0
         sensory = torch.tensor([norm_speed, heading_delta,
                                 self.wall_proximity, coll_val],
-                               device=self.device)
+                               dtype=torch.float32, device=self.device)
         prediction = torch.tensor([self._predicted_speed,
                                    self._predicted_heading_delta,
                                    self._predicted_wall, 0.0],
-                                  device=self.device)
+                                  dtype=torch.float32, device=self.device)
         pc_out = self.pc(sensory_drive=sensory, prediction_drive=prediction)
         pe = pc_out['pe']
         self.speed_pe = float(pe[0])
